Move the language detection to a more appropriate place.
Also, I've allowed an empty string to the locale validator, just to
make sure we don't accidentally block the login/signup/getCurrentSession
APIs due to that.
