+++ client/i18n/i18next.ts
@@ -8,6 +8,7 @@ import {
   TranslationNamespace,
 } from '../../common/i18n'
 import { makeServerUrl } from '../network/server-url'
+import { JsonSessionStorageValue } from '../session-storage'
 
 const isDev = __WEBPACK_ENV.NODE_ENV !== 'production'
 
@@ -19,6 +20,16 @@ const isDev = __WEBPACK_ENV.NODE_ENV !== 'production'
  * https://github.com/i18next/react-i18next/issues/1483#issuecomment-1268455602
  */
 export type TransInterpolation = any
+
+/**
+ * The locale that was reported to us by the user's browser. This locale can be overwritten by
+ * user's explicit choice in the top-links dropdown. We send this locale to the server during
+ * login/signup/getCurrentSession actions.
+ *
+ * NOTE(2Pac): Session storage gets cleared after logging in, so be careful about using this only in
+ * the logged-out pages.
+ */
+export const detectedLocale = new JsonSessionStorageValue<string | undefined>('detectedLocale')
 
 export const languageDetector = new LanguageDetector(null, {
   order: ['navigator'],
