+++ server/lib/i18n/locale-validator.ts
@@ -1,3 +1,3 @@
 import Joi from 'joi'
 
-export const joiLocale = () => Joi.string().max(35)
+export const joiLocale = () => Joi.string().max(35).allow('')
